Webservices

+++ students_app/views.py
@@ -5,14 +5,23 @@ from.models import Students
 from rest_framework.response import Response
 from rest_framework import status
 
-class StudentsList(APIView):
 
+
+class StudentsList(APIView):
+    def authenticate(self, request):
+        username = request.headers.get('Username')
+        password = request.headers.get('Password')
+        return username == 'admin' and password == 'PWD'
     def get(self,request):
+
         students = Students.objects.all()
         serializer= studentSerializers(students,many=True)
         return Response(serializer.data)
 
     def post(self,request):
+        if not self.authenticate(request):
+            return Response({'error': 'Unauthorized'}, status=status.HTTP_401_UNAUTHORIZED)
+
         serializer  = studentSerializers(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -20,6 +29,12 @@ class StudentsList(APIView):
         return Response(serializer.errors,status=status.HTTP_400_BAD_REQUEST)
 
 class StudentByID(APIView):
+
+    def authenticate(self, request):
+        username = request.headers.get('Username')
+        password = request.headers.get('Password')
+        return username == 'admin' and password == 'PWD'
+
     def get_object(self,pk):
         return Students.objects.get(pk=pk)
 
@@ -29,6 +44,10 @@ class StudentByID(APIView):
         return Response(serializer_obj.data)
 
     def put(self,request,pk):
+
+        if not self.authenticate(request):
+            return Response({'error': 'Unauthorized'}, status=status.HTTP_401_UNAUTHORIZED)
+
         students_obj = self.get_object(pk)
         serializer_obj = studentSerializers(students_obj,data=request.data)
         if serializer_obj.is_valid():
@@ -37,8 +56,13 @@ class StudentByID(APIView):
         return Response(serializer_obj.errors,status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self,request,pk):
+
+        if not self.authenticate(request):
+            return Response({'error': 'Unauthorized'}, status=status.HTTP_401_UNAUTHORIZED)
+
         students_obj=self.get_object(pk)
         students_obj.delete()
-        return Response( status=status.HTTP_204_NO_CONTENT)
+        return  Response( status=status.HTTP_204_NO_CONTENT)
+
     def tried(self):
         print('ggggggg')
